refactored files for config.py
