+++ feats/create_h5data.py
@@ -1,6 +1,4 @@
-import csv
-
-from entrainment_config import *
+from entrainment.config import *
 
 SEED=448
 frac_train = 0.8
