+++ models/NED/feat_extract_nopre.py
@@ -8,7 +8,7 @@ import sys
 ###Fix This File: it's not putting all IPU features
 print(sys.path)
 
-from entrainment_config import *
+from entrainment.config import *
 
 np.set_printoptions(threshold=np.inf)
