+++ scripts/activate_virtualenv
@@ -1,4 +1,5 @@
 #!/usr/bin/env bash
+set -eu
 # This script activates the virtual environment for the project
 # creating it if it doesn't exist.
 # Usage: ./activate_virtualenv
@@ -15,5 +16,5 @@ if [[ ! -d "$VENV_DIR" ]]; then
     echo "Thus, we will create it now."
     python3 -m venv "$VENV_DIR"
 fi
-echo $VENV_DIR
+echo $VENV_DIR 
 . "$VENV_DIR"/bin/activate
